move logic to SkyblockStat.kt
